tools/requirements.txt
azure-core==1.32.0
azure-data-tables==12.6.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
isodate==0.7.2
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
